castに成功。

+++ Receiver/py/readfromkafka.py
@@ -1,6 +1,6 @@
 from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import explode
-from pyspark.sql.functions import split, to_binary,encode,decode,unhex,unbase64
+from pyspark.sql.functions import split, to_binary,encode,decode,unhex,unbase64,to_number
 from pyspark.sql.types import DoubleType, FloatType, LongType, StructType,StructField, StringType
 
 spark = SparkSession \
@@ -16,7 +16,7 @@ df = spark \
   .load()
 
 #df2=df.selectExpr("CAST(value AS STRING) as value")
-df2=df.select(decode(df.key,'UTF-8').alias("key"),df.value.cast('STRING').alias("value"))
+df2=df.select(decode(df.key,'UTF-8').cast('FLOAT').alias("key"),df.value.cast('STRING').alias("value"))
 #df.select(df2.value,decode(df.value,'UTF-8'))
 df.show()
 df.printSchema()
